Add "accounting" report
Wikimedia's Accounting department maintains a spreadsheet with data
center assets, as gathered by invoices sent to the Foundation's AP. Each
row represents a single asset, and has fields for the invoice number,
its cost, its PO #, as well as, more importantly, its serial number,
asset tag and procurement task ID. There are currently no plans to move
this to a different system or to a database.

Given this information is manually maintained, and by a different
department and using an entirely different data source (invoices) no
less, discrepancies with our own DCIM data are to be expected. It is
therefore meaningful to compare, and identify typos and other types of
errors on either side.

This report uses the Google Sheets API(!), and performs these checks:
* For a given serial, asset tags match
* For a given serial, procurement tickets match
* No assets exist in Netbox that are not tracked by accounting
* No assets have been invoiced but are not tracked in Netbox

Change-Id: I960241aeaf397e0fe5cfd1474a5d2e78291aa99b