coherence: optimize by reducing database queries
Add a select_related() to test_offline_rack(), that removes a few extra
queries per failed devices.

Add the custom field accessor tricks that we employed in the Accounting
report (and optimized further here). This has also been reported
upstream as #3185, so that we can hopefully avoid repeating this code in
all of our reports.

With today's dataset, this reduces the number of queries from 8336 to a
constant 14, i.e. a ~600x improvement.

Change-Id: I3d80b77802d2ce1eace2d055069426cade0fee33