Remove the oldhardware report
Due to various complications that have to do with how we plan hardware
replacements, this hasn't proved to be as useful as we were expecting it
to. In its current form it's doubtful that is useful to anyone and no
well-defined ideas exist on how it could be made to be useful.

Remove to avoid any distractions or misunderstandings about its use.

Change-Id: I184a5e6fb285865b2b772d5d3b7c72bd9343992c